fix: Ensure untraced blocks suppress spans
Prior to this change, the Net::HTTP instrumentation would defer to the sampling strategy
to determine if it should generate a span, however the sampling strategy
is not aware of the `untraced` block and it will always generate a span
even though the parent span is untraced.

The question here is... should the `untraced` block override the logic in the sampling strategy?

Also... a side effect of checking that the `current_span.recording?` is that the instrumentation
will never be allowed to generate a root span.

Related to https://github.com/open-telemetry/opentelemetry-ruby/discussions/1333