fix: Rename HTTP CONNECT to HTTP SESSION
The current span name that representes `Net::HTTP#start` is a little misleading and should be renamed to `HTTP SESSION` to better represent the span.

According to `Net::HTTP#start` RDoc the `start` method is used to:

> Opens a TCP connection and HTTP session.
> When this method is called with a block, it passes the Net::HTTP object to the block, and closes the TCP connection and HTTP session after the block has been executed.
> When called with a block, it returns the return value of the block; otherwise, it returns self.
<https://ruby-doc.org/stdlib-3.1.2/libdoc/net/http/rdoc/Net/HTTP.html#method-i-start>

These operations are lower level operations that establish SSL connections and initiate an HTTP session,
which is distinct from using the HTTP Protocol `CONNECT` method[^2] which is used to:

> starts two-way communications with the requested resource. It can be used to open a tunnel.
> For example, the CONNECT method can be used to access websites that use SSL (HTTPS).
> The client asks an HTTP Proxy server to tunnel the TCP connection to the desired destination.
> The server then proceeds to make the connection on behalf of the client.
> Once the connection has been established by the server, the Proxy server continues to proxy the TCP stream to and from the client.
> `CONNECT www.example.com:443 HTTP/1.1`

<https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/CONNECT>

`Net::HTTP` does use the `HTTP CONNECT` method but _only_ when the `Net::HTTP` object is configured to use a proxy:

https://github.com/ruby/net-http/blob/196f3d70f9ce154dc63a86a2cdbf1aec4df4982e/lib/net/http.rb#L1031

If we want to measure the time it takes to complete an `HTTP CONNECT` method request the code should create a separate span _or_ check
the `proxy?` predicate.