fix: only take the first item in a comma-separated list for pg attrs
In a recent upstream PR[^1], a change was made that introduced the
possibility of the `conninfo_hash` containing a comma-separated list if
the hostname we connect to resolves to multiple IP addresses.

In the tests, we say to connect to `localhost`: and sure enough, that
resolves to both `127.0.0.1` and `::1`. :facepalm:

I considered simply relaxing the validation here, but we'd break
semantic conventions: `net.peer.port` must be an integer, not a string -
so we can't just pass through the unmodified values. The values will
always be duplicate, comma-separated strings however: the actual
difference shows up in `net.peer.ip`.

Anyways, we add a helper function that tries to be smart about not doing
more work than is necessary.