ci: split windows/mac and truffleruby builds to scheduled canary
We want to move windows/mac builds (as well as all truffleruby builds)
to a nightly canary build. We care about them, but much less so than
linux + MRI/JRuby. There are periodic segfaults on the truffleruby
builds (which aren't clearly related to our code), and the Mac/Windows
builds can take awhile to run due to limited runner availability.

The general plan is to watch and see how much usage increases for these
platforms (if at all), and if they do we can add them back in to the
regular CI matrix.

The CI configuration is getting unwieldy again, so I may take another
pass at simplifying the configuration in the future.

cc #103