ActionPack - Add source code attributes, change enable_recognize_route default from false to true (#170)
* feat: Set code.namespace & code.function in ActionPack instrumentation

This is in line with how the OpenTelemetry Semantic Conventions specify
referencing source code:

https://opentelemetry.io/docs/reference/specification/trace/semantic_conventions/span-general/#source-code-attributes

These attributes will identify the Controller and Action called when
servicing a request. These attributes are what make up the
:controller_action span naming for the ActionPack instrumentation.

Controller#action was the default Rack span name when using the
ActionPack instrumentation, but is now configurable. When using the
:rails_route span naming you lose the ability to know what controller
and action a route would call. Rather than add a new attribute with the
Controller#action format it was preferred to follow the existing
Semantic Convention.

* refactor: Change default option for ActionPack :enable_recognize_route to true

The default span naming for ActionPack instrumentation uses the
recognize route feature, leaving :enable_recognize_route with the
default of false will have negligible performance gains given the
recognized route is memoized in the instrumentation.

Co-authored-by: Ariel Valentin <arielvalentin@users.noreply.github.com>