fix: bump rails instrumentation dependency on action_pack instrumentation
The change in #157 won't get picked up by users of
`opentelemetry-instrumentation-rails` unless they explicitly do a
`bundle update` or install the gem anew.

It's definitely not like, the _most_ critical of fixes, but it seems to
merit inclusion so that people are not confused as to why the documented
options aren't working as claimed. :tada: