ci: Break out datastores into separate builds
Builds have been unstable lately due to failing services starting up and SEGFAULT in kafka gems.

This change is an attempt at splitting up builds by datastore so they fail earlier and make it possible to rerun if necessary.

The disadvantaage of this approach is that it will require additional runners to parallelize the builds.