ci: disable a vexatious rubocop rule, and correct a sorting
Rubocop has been failing our canary builds every night for a longer
period of time than I'm willing to admit to in a public forum.

One of these failures was because we weren't strictly ensuring that
development dependencies were listed in the Gemfile as opposed to the
gemspec. I really think that's a little over-pedantic, so I've turned
it off with the support of my esteemed colleage, @arielvalentin.

The other failure we fix is related to the sorting of dependencies
within the gemspec. Mind you, I think this failure is also
overly-pedantic and I would be inclined to disable it. However, rubocop
was able to autofix it with no effort on my part, and so I've kept the fix.