feat: Use Rack::Events for instrumentation (#342)
* feat: Use Rack::Events for instrumentation

This is an alternate implementation of the Rack instrumentation that leverages the [`Rack::Events` API](https://www.rubydoc.info/gems/rack/Rack/Events) instead of a custom`Rack::Middleware`.

Why am I suggesting we change this instrumentation? At GitHub we leverage `Rack::BodyProxy` to write Rack Request logs after the request is complete, however Rack span is already finished and its related `Context` has already been detached. This means we are not able to correlate request logs to our traces.

The advantage of using `Rack::Events` is that handlers are triggered during difference stages of a request, including for deferred operations like [`Rack::BodyProxy`](https://www.rubydoc.info/gems/rack/Rack/BodyProxy), as opposed to middlewares who _only_ are invoked inline.

The disadvantage of this API is that it makes managing the request more difficult and we have to track the Context positions to detach in the `Rack::Env`.

This implementation will be released along side the existing instrumtation to give users the option to use the middleware instead of the `Rack::Events` handler until we are able to run this in some of our heavy production workloads.

Fixes https://github.com/open-telemetry/opentelemetry-ruby-contrib/issues/341
Related https://github.com/DataDog/dd-trace-rb/pull/1746

* squash: additional feature parity

* squash: add allowed response headers

* squash: url quantization

* squash: Now with new Lemon Scented response headers

* squash: we are now at parity

* squash: use instrumetation config

* squash: Use declarative config options

* squash: fix bad refactoring

* convert proxy span to an event

* refactor: move configurations to instrumentation install

* squash: add test converage

* squash: make response headers a little more resilient

* squash: Ensures event middleware will not cause the application to crash when it encounters errors

* squash: fix linter error

* feat: Add middleware args helper for ActionPack and Sinatra

* fix: test case

* fix: Rack Events is autoloaded so if the parent module is present so are submodules

* fix: More precise error handling

* fix: Ensure config is cleared/setup during installation

* fix: Sinatra 1.4 compatability

* fix: bad merge

* fix: invalid responses in test case

* squash: Added a few more test cases