Grape Instrumentation: handle status code symbol
ever since grape 1.7.0 grape endpoints can return status codes as
rack status code symbols, these then are converted to integers by the
error middleware. We need to handle. In event_handler we make the assumption that
exception status would be an integer.
Now we correctly convert the status code from a symbol to an integer.
 I basically copied `Rack::Utils` implementation to prevent
 adding rack as a dependency.

The only change is if the status_code is not valid the status code is
assumed to be 0.