fix: Grape Instrumentation handle status code symbol (#448)
Since grape 1.7.0 grape endpoints can return status codes as rack
status code symbols. The symbols are converted to integers by the
Grape error middleware.

This performs the same conversion with `Rack::Utils`. Rack is already
an implicit runtime dependency (and should remain implicit) of the
instrumentation, but has now been added as an explicit development
dependency for the tests.

If the status_code is not valid the status code is assumed to be 0.