feat!: Separate logical MySQL host from connected
MySQL server instances have a logical host name and an instance node names.

This is particularly useful when running in clustered mode where a logical hostname
is used to connect to a cluster but you also want to record what specific node is executing your query.

Trilogy exposes the instance node name via `@connected_host` as well as
the configured `host` name in its `connection_options`.

This breaking change ensures that the `net.peer.name` is always set to the logical host name,
while the individual instance name is being recorded in a separate attribute.