feat: Ensure Sidekiq TracerMiddleware is early in the chain
This change will ensure the middleware is put at the front of the Client
and Server middleware chains, by default. This is a change from the
existing behavior which will add them to the end of the chain, and that
position is determined entirely by when the SDK is configured. In a
Rails app, that usually means it depends on the name of the
`config/initializers/*.rb` file `OpenTelemetry::SDK.configure` is called
from.

It is still possible that other middlewares end up in front of the
TracerMiddleware as other Sidekiq extensions/feature might
`chain.prepend` after our configure block runs. There's just not much we
can do about that. But at least _trying_ to be early in the chain is a
better, more predictable default than wherever we happen to be called.