fix!: Drop DelayedJob ActiveRecord in Tests
The DelayedJob tests suite relied on the ActiveRecord Backend
gem for testing, however incompatabilities with ActiveRecord/Rails 7.1 means would
result in failures in the test suite.

It is not the responsibility of OTel Ruby Instrumentation to ensure compatability
of 3rd party gems or ensure our test suite to come up with workarounds for bugs.

For this reason I have rewritten the test suite to use the Test backend that Delayed Job
used to test the upstream gem.

This does introduce one change, the `created_at` attribute, which is _specific_ to the ActiveRecord backend
will _no longer_ be emitted. Users that need this event should pin to earlier versions of the gem.

There is a risk here that the upstream DelayedJob gem will no longer be included in future releases,
but I think that is less likely than to run into incompatabilities with future Rails versions.