refactor: Default to Proxy Tracer in instrumentations
While working on https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/677,
I ran into an issue where the base instrumentation tracer was not being automatically upgraded
after the SDK initialization was complete.

I am not sure why this was implemented this way but I figured other instrumentations would need a reference to a proxy tracer in the future.

https://github.com/open-telemetry/opentelemetry-ruby/pull/671