Add support for ActionController::Live
This commit adds a patch to ensure `ActionController::Live` actions,
which are processed in a new thread, are instrumented as a child span
in the current trace.

This is a patch rather than an `ActiveSupport::Notifications` handler.
Although the `action_controller.process_action` event is processed in
the new thread, we don't have a reference to the controller or anything
to indicate that it is an `ActionController::Live` controller.

The patch needs to be included rather than prepended, as
`ActionController::Live` is a module that is included in each controller
as required by the host app, rather than a class to inherit from. As
such, `include` is the only way to get our `#process_action` method into
the right place.

We need to manually unset the `:__opentelemetry_context_storage__`
thread local, as this gets copied across from the parent thread and
points to the same stack array. Calling `#clear` would modify that
shared stack array, so we need to unset it instead.