fix: Relax otel common gem constraints
Rolling out changes to opentelemetry-common is made difficult because many of the instrumentation gems that rely on it have to all receive minor version bumps.

Relaxing constraints allows for individual instrumentations to be able to use new features in the common gem without requiring a minor dependency bump of unrelated instrumentations that share that dependency.