feat: Add after_request hook to Rack tracer middleware
The intention here is to allow the user to be able to modify the span
attributes using details about the response. This PR adds an
`after_request` hook to the configuration options for the tracer
middleware so that the user may supply a method which can interact
with the span attributes.