fix: compatibility with Faraday v1 (#1032)
Fixes https://github.com/open-telemetry/opentelemetry-ruby-contrib/issues/1031

Faraday >= 1 changes where the #adapter method gets called, affecting
the RackBuilder patch in this gem.

Without this change, the tracer middleware gets added to the beginning
of the stack instead of the end, and explicitly adding the middleware
would cause it to get added twice.

This change patches `Connection#initialize` for Faraday >= 1 to preserve
the behavior we had with Faraday < 1. This is similar to the dd-trace-rb
change made in https://github.com/DataDog/dd-trace-rb/pull/906.

There is some followup work here to explore moving the middleware to the
beginning of the stack instead of the end, but we can work on that
separately.