feat: optional rack instrumentation install by grape instrumentation
In some circumstances we may want to defer the installation of the rack
instrumentation that is orchestrated during the grape instrumentation
installation. For example, we may want to customise the installation of
the rack instrumentation or we can rely on the installation being
orchestrated by another framework's instrumentation (such as when Grape
is mounted with Rails).

This change allows for Grape to skip the installation of the Rack
instrumentation so that it can manually installed at a later time.