chore: Update instrumentation/aws_sdk CODEOWNERS (#1191)
Update instrumentation/aws_sdk CODEOWNERS

The wonderful @jterapin and @alextwoods are members of the team at AWS
who own the AWS SDK for Ruby. They've contributed a few changes to the
aws_sdk instrumentation for enhanced compatibility with v2 and v3.

This commit adds them to the codeowners list for the
opentelemetry-instrumentation-aws_sdk gem.