test: Add less restrictive gem spec for mongo gem (#1537)
* Add less restrictive gem spec for mongo gem

It looks like latest features (including jruby 10 enablement) will require the
latest minor release of the mongo gem.

Specifically:

The latest mongo gem relaxes the range for the ruby-bson gem. This would allow
the uptake of https://github.com/mongodb/bson-ruby/pull/353 if it is accepted.

* Update test expectations based on latest mongo gem

This commit updates the test expectation for an updated message string. The same
behavior is observbed, the message has just been updated.

This commit also updates construction of a test mongodb client with a newly
required value for auth_source when building a simple client.