fix: Fixing missing OpenTelemetry::Context detach on Excon instrumentation
When a span is not recorded, for example when you are sampling traces,
the OpenTelemetry::Context attached on `request_call` is never detached.
This happens because `handle_response` is doing nothing if
`span.recording?` is false.

Here is example of the error that the missing detach generated:
```
ERROR -- : OpenTelemetry error: calls to detach should match corresponding calls to attach.
```