fix: Fixing missing OpenTelemetry::Context detach on Excon instrumentation (#1725)
* fix: Fixing missing OpenTelemetry::Context detach on Excon instrumentation

When a span is not recorded, for example when you are sampling traces,
the OpenTelemetry::Context attached on `request_call` is never detached.
This happens because `handle_response` is doing nothing if
`span.recording?` is false.

Here is example of the error that the missing detach generated:
```
ERROR -- : OpenTelemetry error: calls to detach should match corresponding calls to attach.
```

* Fixing tests following action_pack strategy to change the sampler

* Improving datum handling

---------

Co-authored-by: Kayla Reopelle <87386821+kaylareopelle@users.noreply.github.com>