chore: Add component owners workflow, documentation (#1771)
* chore: Add component owners workflow

Add the component owners workflow to automatically assign owners to
pull requests when changes are proposed for their gem.

This is blocked until we can make sure all component owners are members
of the OpenTelemetry organization on GitHub and are added as members of
the ruby-contrib-triagers team.

* ci: Add Triager, triagers to known words

* Apply suggestions from code review

Co-authored-by: Ariel Valentin <arielvalentin@users.noreply.github.com>

* Add factory_bot to component owners

* Keep wildcard in codeowners

This intends to continue our practice of having every maintainer
and approver directly assigned to
PRs

* Update CODEOWNERS

---------

Co-authored-by: Ariel Valentin <arielvalentin@users.noreply.github.com>