fix: use a unique consumer group in tests
The tests for this instrumentation are failing on racecar 2.10+. This is
due to an issue caused by reuse of consumer group ids (derived from the
class name in racecar unless specified) in tests.
Switching to use a unique consumer group fixes the tests.

In Racecar 2.10 a [change](https://github.com/zendesk/racecar/pull/339)
was intrdouced to add a RebalanceListener by default to every consumer.

Now, when a consumer leaves a group, rdkafka's
rebalance protocol waits for acknowledgment from the listener before
completing the leave.

When the test's consumer closes and the next test's consumer tries to
join the same group ID, Kafka's group coordinator now waits for
the rebalance protocol to complete, which takes longer than before due
to the new listener-based coordination.